requirements.txt
accelerate==0.17.1
albumentations==1.3.0
antlr4-python3-runtime==4.9.3
certifi==2022.12.7
charset-normalizer==3.1.0
dominate==2.7.0
fvcore==0.1.5.post20221221
idna==3.4
imageio==2.26.0
importlib-metadata==6.1.0
iopath==0.1.10
joblib==1.2.0
lpips==0.1.4
networkx==2.6.3
numpy==1.21.6
omegaconf==2.3.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.0
Pillow==9.4.0
pip==22.3.1
portalocker==2.7.0
prefetch-generator==1.0.3
protobuf==3.20.3
psutil==5.9.4
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.28.2
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
setuptools==65.6.3
tabulate==0.9.0
tensorboardX==2.6
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.12.1+cu102
torch-ema==0.3
torchaudio==0.12.1+cu102
torchscan==0.1.2
torchvision==0.13.1+cu102
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
wheel==0.38.4
yacs==0.1.8
zipp==3.15.0
